refactor(windows): extract modules from `gui` module
